-Bumped the version of some libraries -Added location services to the build.gradle -Changed the viewModels to use the SavedStateViewModelFactory -Added code to update the user provider on login -Fixed bug where the user ( admin ) could click an item with zero likes to see them -Fixed bug preventing the coloring of the like heart properly -Changed onDocumentAdded from private to protected -Added alertdialog questioning if the user ( admin ) really wishes to delete the like -Added alertdialog questioning if the user really wishes to delete the message -Added alertdialog questioning if the user really wishes to delete its like -Fixed lastSeen bug in the UserAdapter -Added scopes to the AuthActivityResultContract -Added UnpublishIfNotFoundInFirebase option to the reload products screen -Added fetched field in the messages entity -Created inPublishedSource compound field in the Product -Added providerID to the User entity -Updated the viewModels to use the savedStateHandle parameter -Fixed workers bug that prevented them from beginning -Added code the the splashActivity to retrieve the remote config -Added geolocation in the home fragment. If the user is in Brazil, it will show the Iluria Store. The Etsy will be implemented afer -Bumped the gradle version from 6.6.1 to 6.7 -Created NewMessagesService to notify the admin when new messges arrive

+++ app/src/main/java/app/web/diegoflassa_site/littledropsofrain/adapters/MessageAdapter.kt
@@ -1,5 +1,6 @@
 package app.web.diegoflassa_site.littledropsofrain.adapters
 
+import android.app.AlertDialog
 import android.content.Context
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -90,13 +91,13 @@ open class MessageAdapter(
             itemView.setOnClickListener {
                 replyMessage(it, message!!)
             }
+            binding.msgImage.visibility = View.GONE
+            binding.btnViewAsNotification.visibility = View.GONE
 
             // Click listener
             itemView.setOnClickListener { listener?.onMessageSelected(snapshot) }
             when (MessageType.valueOf(message?.type?.toUpperCase(Locale.ROOT)!!)) {
                 MessageType.MESSAGE -> {
-                    binding.msgImage.visibility = View.GONE
-                    binding.btnViewAsNotification.visibility = View.GONE
                     when {
                         message.emailSender == LoggedUser.userLiveData.value?.email -> {
                             itemView.background = ContextCompat.getDrawable(
@@ -128,7 +129,7 @@ open class MessageAdapter(
                     )
                 }
                 MessageType.UNKNOWN -> {
-                    // DO nothing
+                    // Do nothing
                 }
             }
 
@@ -167,7 +168,20 @@ open class MessageAdapter(
                 )
             )
             binding.btnDelete.setOnClickListener {
-                MessageDao.delete(message)
+                val builder = AlertDialog.Builder(mContext)
+                builder.setMessage(mContext.getString(R.string.remove_message_confirmation))
+                    .setCancelable(false)
+                    .setPositiveButton(mContext.getString(R.string.yes)) { _, _ ->
+                        ioScope.launch {
+                            MessageDao.delete(message)
+                        }
+                    }
+                    .setNegativeButton(mContext.getString(R.string.no)) { dialog, _ ->
+                        // Dismiss the dialog
+                        dialog.dismiss()
+                    }
+                val alert = builder.create()
+                alert.show()
             }
         }
